Access modificators are fixed. Lab works task are moved to the separate project.
