+++ CrossplatformsTasks.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32630.192
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrossplatformsTasks", "CrossplatformsTasks.csproj", "{92CC9604-7B70-4974-88F6-0BC4198349D9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrossplatformsTasks", "MainPart\CrossplatformsTasks.csproj", "{92CC9604-7B70-4974-88F6-0BC4198349D9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lab123", "Lab123\Lab123.csproj", "{7D26A894-58F0-4D6F-B8A2-B09B98AD8557}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{92CC9604-7B70-4974-88F6-0BC4198349D9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{92CC9604-7B70-4974-88F6-0BC4198349D9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{92CC9604-7B70-4974-88F6-0BC4198349D9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7D26A894-58F0-4D6F-B8A2-B09B98AD8557}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7D26A894-58F0-4D6F-B8A2-B09B98AD8557}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7D26A894-58F0-4D6F-B8A2-B09B98AD8557}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7D26A894-58F0-4D6F-B8A2-B09B98AD8557}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
