+++ Lab123/Lab123.csproj
@@ -1,14 +1,15 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-	<Folder Include="Lab1Task\" />
-  </ItemGroup>
-
+	<ItemGroup>
+		<Folder Include="Lab1Task\" />
+		<Folder Include="Lab2Task\" />
+		<Folder Include="Lab3Task\" />
+	</ItemGroup>
+	
 </Project>
