+++ Lab123/Lab1Task/CommunicationMessagesLab1.cs
@@ -1,6 +1,6 @@
-﻿namespace CrossplatformsTasks.Lab2Task
+﻿namespace Lab123.Lab1Task
 {
-    public static class CommunicationMessagesLab2
+    public static class CommunicationMessagesLab1
     {
         public const string InputPathToFileMessage = "\nPlease input a path to your text file: ";
         public const string TryAgainMessage = "If you want to try again please input \"1\":";
