+++ Lab123/Lab1Task/ExceptionMessagesLab1.cs
@@ -1,10 +1,10 @@
-﻿namespace CrossplatformsTasks.Lab2Task
+﻿namespace Lab123.Lab1Task
 {
-    public static class ExceptionMessagesLab2
+    public static class ExceptionMessagesLab1
     {
         public const string InvalidPathToFileMessage = "\nYou specified the wrong path to your file with the input data or the file does not exist.";
         public const string NotNumberInputMessage = "\nInput value in the file was not a number.";
         public const string ExceedingLimitsNumberMessage = "\nInput value in the file exceeds the limits: only numbers from 2 to 100 are allowed.";
-        public const string DifferentNumberOfStairsMessage = "\nNumber of input values in the second line of the file is different from the number in the first line given.";
+
     }
 }
