+++ Lab123/Lab2Task/CommunicationMessagesLab2.cs
@@ -1,6 +1,6 @@
-﻿namespace CrossplatformsTasks.Lab3Task
+﻿namespace Lab123.Lab2Task
 {
-    public static class CommunicationMessagesLab3
+    public static class CommunicationMessagesLab2
     {
         public const string InputPathToFileMessage = "\nPlease input a path to your text file: ";
         public const string TryAgainMessage = "If you want to try again please input \"1\":";
