+++ Lab123/Lab3Task/CommunicationMessagesLab3.cs
@@ -1,6 +1,6 @@
-﻿namespace CrossplatformsTasks.Lab1Task
+﻿namespace Lab123.Lab3Task
 {
-    public static class CommunicationMessagesLab1
+    public static class CommunicationMessagesLab3
     {
         public const string InputPathToFileMessage = "\nPlease input a path to your text file: ";
         public const string TryAgainMessage = "If you want to try again please input \"1\":";
