+++ MainPart/Program.cs
@@ -1,0 +1,7 @@
+﻿using Lab123.Lab1Task;
+using Lab123.Lab2Task;
+using Lab123.Lab3Task;
+
+//Lab1.StartNewDialogForLab1();
+//Lab2.StartNewDialogForLab2();
+Lab3.StartNewDialogForLab3();
