--- Program.cs
@@ -1,7 +1,0 @@
-﻿using CrossplatformsTasks.Lab1Task;
-using CrossplatformsTasks.Lab2Task;
-using CrossplatformsTasks.Lab3Task;
-
-//Lab1.StartNewDialogForLab1();
-//Lab2.StartNewDialogForLab2();
-Lab3.StartNewDialogForLab3();
